Controller now updating the QServer with Q values
Also moved a bunch of files.

I think most of the worker side is done...
Some work needs to be done to transfer results after completion.
These results would be things like cumulative result and so on.

Then I just need to add some export functionality to the server
and to work on the server being able to run multiple experiments.
And to launch workers from the server.