Update QServer to have Results endpoint and track an experiment
The server State now tracks experiment data like number
the number of finished agents and number of agents in the experiment
and if the experiment is finished or not.

And end point has been added for agents to send their final metrics
to be aggregated by the server in an Experiment results object

An endpoint to get the state has been added to be able to access
results and the status of training.